--- src/main/java/org/devkor/apu/saerok_server/domain/dex/bird/contract/HasBodyLength.java
@@ -1,5 +1,0 @@
-package org.devkor.apu.saerok_server.domain.dex.bird.contract;
-
-public interface HasBodyLength {
-    Double getBodyLengthCm();
-}
