+++ src/main/java/org/devkor/apu/saerok_server/domain/dex/bird/domain/contract/HasBodyLength.java
@@ -1,0 +1,5 @@
+package org.devkor.apu.saerok_server.domain.dex.bird.domain.contract;
+
+public interface HasBodyLength {
+    Double getBodyLengthCm();
+}
