+++ src/main/java/org/devkor/apu/saerok_server/domain/dex/bird/domain/service/SizeCategoryService.java
@@ -1,0 +1,31 @@
+package org.devkor.apu.saerok_server.domain.dex.bird.domain.service;
+
+import lombok.RequiredArgsConstructor;
+import org.devkor.apu.saerok_server.domain.dex.bird.domain.contract.HasBodyLength;
+import org.devkor.apu.saerok_server.domain.dex.bird.domain.dto.SizeCategoryDto;
+import org.devkor.apu.saerok_server.global.config.SizeCategoryRulesConfig;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class SizeCategoryService {
+
+    private final SizeCategoryRulesConfig rules;
+
+    public SizeCategoryDto getSizeCategory(HasBodyLength item) {
+        Double bodyLengthCm = item.getBodyLengthCm();
+
+        if (bodyLengthCm == null) {
+            return SizeCategoryDto.Empty();
+        }
+
+        for (SizeCategoryRulesConfig.Boundary boundary : rules.getBoundaries()) {
+            if (boundary.getMaxCm() == null || bodyLengthCm <= boundary.getMaxCm()) {
+                String category = boundary.getCategory();
+                return new SizeCategoryDto(category, rules.getLabels().get(category));
+            }
+        }
+
+        return SizeCategoryDto.Empty();
+    }
+}
