--- src/main/java/org/devkor/apu/saerok_server/domain/dex/bird/dto/service/SizeCategoryDto.java
@@ -1,15 +1,0 @@
-package org.devkor.apu.saerok_server.domain.dex.bird.dto.service;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class SizeCategoryDto {
-    private String code; // ex: "small"
-    private String label; // ex: "참새 크기"
-
-    public static SizeCategoryDto Empty() {
-        return new SizeCategoryDto(null, null);
-    }
-}
