+++ src/main/java/org/devkor/apu/saerok_server/domain/dex/residency/entity/BirdResidency.java
@@ -1,7 +1,7 @@
 package org.devkor.apu.saerok_server.domain.dex.residency.entity;
 
 import jakarta.persistence.*;
-import org.devkor.apu.saerok_server.domain.dex.bird.entity.Bird;
+import org.devkor.apu.saerok_server.domain.dex.bird.domain.entity.Bird;
 import org.devkor.apu.saerok_server.global.entity.Auditable;
 
 @Entity
